+++ app/src/main/java/com/grace/inventory/data/products.kt
@@ -30,7 +30,31 @@ import java.time.temporal.TemporalAmount
 //)
 
 
-
+//
+//data class Product(
+//    var productId: String = "",
+//    var productName: String = "",
+//    var totalPrice: Double = 0.0,
+//    var totalAmount: Double = 0.0,
+//    var date: String = "",
+//    var timestamp: Long = 0L,
+//    var expenseId: String = "",
+//    var saleId: String = "",
+//    var transactionId: String = "",
+//    var quantitySold: Int = 0,         // 🔁 changed from Int → Long
+//    var tag: String = "",
+//    val quantity: Long = 0,             // 🔁 changed from Int → Long
+//    val sellingPrice: Double = 0.0
+//)
+//
+//data class SaleItem(
+//    val productId: String = "",
+//    val productName: String = "",
+//    val quantitySold: Int = 0,         // 🔁 changed from Int → Long
+//    val totalAmount: Int = 0,
+//    val quantity: Long = 0,             // 🔁 changed from Int → Long
+//    val sellingPrice: Double = 0.0
+//)
 data class Product(
     var productId: String = "",
     var productName: String = "",
@@ -41,17 +65,46 @@ data class Product(
     var expenseId: String = "",
     var saleId: String = "",
     var transactionId: String = "",
-    var quantitySold: Int = 0,         // 🔁 changed from Int → Long
+    var quantitySold: Long = 0L,         // Changed from Int to Long
     var tag: String = "",
-    val quantity: Long = 0,             // 🔁 changed from Int → Long
+    val quantity: Long = 0L,             // Already Long
     val sellingPrice: Double = 0.0
 )
 
 data class SaleItem(
     val productId: String = "",
     val productName: String = "",
-    val quantitySold: Int = 0,         // 🔁 changed from Int → Long
-    val totalAmount: Int = 0,
-    val quantity: Long = 0,             // 🔁 changed from Int → Long
+    val quantitySold: Long = 0L,         // Changed from Int to Long
+    val totalAmount: Long = 0L,          // Changed from Int to Long
+    val quantity: Long = 0L,             // Already Long
     val sellingPrice: Double = 0.0
 )
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
